+++ src/reuse/_comment.py
@@ -350,6 +350,15 @@ class LispCommentStyle(CommentStyle):
     INDENT_AFTER_SINGLE = " "
 
 
+class M4CommentStyle(CommentStyle):
+    """M4 (autoconf) comment style."""
+
+    _shorthand = "m4"
+
+    SINGLE_LINE = "dnl"
+    INDENT_AFTER_SINGLE = " "
+
+
 class MlCommentStyle(CommentStyle):
     """ML comment style."""
 
@@ -439,6 +448,7 @@ EXTENSION_COMMENT_STYLE_MAP = {
     ".lisp": LispCommentStyle,
     ".lsp": LispCommentStyle,
     ".lua": HaskellCommentStyle,
+    ".m4": M4CommentStyle,
     ".markdown": HtmlCommentStyle,
     ".md": HtmlCommentStyle,
     ".ml": MlCommentStyle,
@@ -459,6 +469,8 @@ EXTENSION_COMMENT_STYLE_MAP = {
     ".py": PythonCommentStyle,
     ".pyi": PythonCommentStyle,
     ".pyw": PythonCommentStyle,
+    ".pyx": PythonCommentStyle,
+    ".pxd": PythonCommentStyle,
     ".qbs": CCommentStyle,
     ".qml": CCommentStyle,
     ".rake": PythonCommentStyle,
@@ -467,9 +479,11 @@ EXTENSION_COMMENT_STYLE_MAP = {
     ".rbx": PythonCommentStyle,
     ".rs": CCommentStyle,
     ".rss": HtmlCommentStyle,
+    ".sass": CssCommentStyle,
     ".scala": PythonCommentStyle,
     ".scpt": AppleScriptCommentStyle,
     ".scptd": AppleScriptCommentStyle,
+    ".scss": CssCommentStyle,
     ".sh": PythonCommentStyle,
     ".sml": MlCommentStyle,
     ".sql": HaskellCommentStyle,
@@ -481,24 +495,28 @@ EXTENSION_COMMENT_STYLE_MAP = {
     ".vala": CCommentStyle,
     ".xml": HtmlCommentStyle,
     ".xsh": PythonCommentStyle,
+    ".xsl": PythonCommentStyle,
     ".yaml": PythonCommentStyle,
     ".yml": PythonCommentStyle,
     ".zsh": PythonCommentStyle,
 }
 
 FILENAME_COMMENT_STYLE_MAP = {
+    ".editorconfig": PythonCommentStyle,
     ".gitattributes": PythonCommentStyle,
     ".gitignore": PythonCommentStyle,
     ".gitmodules": PythonCommentStyle,
-    ".editorconfig": PythonCommentStyle,
+    ".mailmap": PythonCommentStyle,
     ".pylintrc": PythonCommentStyle,
-    "Dockerfile": PythonCommentStyle,
     "CMakeLists.txt": PythonCommentStyle,
+    "Dockerfile": PythonCommentStyle,
     "Makefile": PythonCommentStyle,
+    "Makefile.am": PythonCommentStyle,
     "Manifest.in": PythonCommentStyle,
+    "ROOT": MlCommentStyle,
+    "configure.ac": M4CommentStyle,
     "manifest": PythonCommentStyle,  # used by cdist
     "requirements.txt": PythonCommentStyle,
-    "ROOT": MlCommentStyle,
     "setup.cfg": PythonCommentStyle,
 }